Split acquisition module to internal vs shared, avoid importing from package modules when possible.
